# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [todo, in_progress, done]
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
  /habits:
    get:
      operationId: listHabits
      tags: [habits]
      summary: List all habits
      responses:
        "200":
          description: List of habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Habit"
    post:
      operationId: createHabit
      tags: [habits]
      summary: Create a habit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHabitBody"
      responses:
        "201":
          description: Created habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [low, medium, high]
        projectId:
          type: integer
          nullable: true
        clientId:
          type: integer
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - status
        - priority
        - createdAt
        - updatedAt
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [low, medium, high]
        projectId:
          type: integer
          nullable: true
        clientId:
          type: integer
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
      required:
        - title
        - status
        - priority
...
    Habit:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        frequency:
          type: string
        streak:
          type: integer
        bestStreak:
          type: integer
        totalCompletions:
          type: integer
        color:
          type: string
        icon:
          type: string
        lastCompleted:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - frequency
        - streak
        - bestStreak
        - totalCompletions
        - color
        - icon
        - createdAt
    CreateHabitBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        frequency:
          type: string
        color:
          type: string
        icon:
          type: string
      required:
        - name
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          nullable: true
        company:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        budget:
          type: string
          nullable: true
        service:
          type: string
          nullable: true
        source:
          type: string
        score:
          type: string
          enum: [hot, warm, cold]
        stage:
          type: string
          enum: [new, contacted, qualified, proposal, negotiation, won, lost]
        notes:
          type: string
          nullable: true
        lastContactedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, source, score, stage, createdAt, updatedAt]
    CreateLeadBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        company:
          type: string
        phone:
          type: string
        budget:
          type: string
        service:
          type: string
        source:
          type: string
        score:
          type: string
        stage:
          type: string
        notes:
          type: string
      required: [name]
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        clientId:
          type: integer
          nullable: true
        projectId:
          type: integer
          nullable: true
        amount:
          type: string
        currency:
          type: string
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        dueDate:
          type: string
          format: date-time
          nullable: true
        paidDate:
          type: string
          format: date-time
          nullable: true
        items:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, invoiceNumber, amount, currency, status, createdAt, updatedAt]
    CreateInvoiceBody:
      type: object
      properties:
        invoiceNumber:
          type: string
        clientId:
          type: integer
        projectId:
          type: integer
        amount:
          type: string
        currency:
          type: string
        status:
          type: string
        dueDate:
          type: string
          format: date-time
        items:
          type: string
        notes:
          type: string
      required: [invoiceNumber, amount]
...